+++ tools/corda-runtime-gradle-plugin/src/smokeTest/kotlin/net/corda/gradle/plugin/SmokeTestBase.kt
@@ -1,7 +1,5 @@
 package net.corda.gradle.plugin
 
-import kotlinx.coroutines.runBlocking
-import net.corda.restclient.CordaRestClient
 import org.gradle.testkit.runner.BuildResult
 import org.gradle.testkit.runner.BuildTask
 import org.gradle.testkit.runner.GradleRunner
@@ -11,89 +9,16 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.io.TempDir
 import java.io.File
 import java.io.Writer
-import java.net.URI
-import java.time.Duration
-import java.time.Instant
-import java.util.concurrent.TimeUnit
 
-// https://docs.gradle.org/current/userguide/test_kit.html
 abstract class SmokeTestBase {
-    companion object {
-        private val targetUrl = URI("https://localhost:8888")
-        private const val USER = "admin"
-        private const val PASSWORD = "admin"
-
-        @JvmStatic
-        protected val restClient = CordaRestClient.createHttpClient(targetUrl, USER, PASSWORD, insecure = true)
-
-        private val composeFile = File(this::class.java.getResource("/config/combined-worker-compose.yml")!!.toURI())
-        private const val CORDA_RUNTIME_VERSION_STABLE = "5.3.0.0-HC01"
-        private val testEnvCordaImageTag = System.getenv("CORDA_IMAGE_TAG") ?: CORDA_RUNTIME_VERSION_STABLE
-
-        @JvmStatic
-        protected fun waitUntilRestOrThrow(timeout: Duration = Duration.ofSeconds(120), throwError: Boolean = true) {
-            val start = Instant.now()
-            while (true) {
-                try {
-                    restClient.helloRestClient.getHelloGetprotocolversion()
-                    return
-                } catch (e: Exception) {
-                    if (Duration.between(start, Instant.now()) < timeout) {
-                        Thread.sleep(10 * 1000)
-                        continue
-                    }
-                    if (throwError) {
-                        throw IllegalStateException("Failed to connect to Corda REST API", e)
-                    } else return
-                }
-            }
-        }
-
-        private fun runComposeProjectCommand(vararg args: String): Process {
-            val composeCommandBase = listOf(
-                "docker",
-                "compose",
-                "-f",
-                composeFile.absolutePath,
-                "-p",
-                "corda-cluster"
-            )
-            val cmd = composeCommandBase + args.toList()
-            val cordaProcessBuilder = ProcessBuilder(cmd)
-            // Get preTest image tag from the pipeline, or fallback to stable version
-            cordaProcessBuilder.environment()["CORDA_RUNTIME_VERSION"] = testEnvCordaImageTag
-            cordaProcessBuilder.redirectErrorStream(true)
-            val process = cordaProcessBuilder.start()
-            process.inputStream.transferTo(System.out)
-
-            return process
-        }
-
-        @JvmStatic
-        fun startCompose(wait: Boolean = true) {
-            val cordaProcess = runComposeProjectCommand("up", "--pull", "missing", "--quiet-pull", "--detach")
-
-            val hasExited = cordaProcess.waitFor(10, TimeUnit.SECONDS)
-            if (cordaProcess.exitValue() != 0 || !hasExited) {
-                throw IllegalStateException("Failed to start Corda cluster using docker compose")
-            }
-
-            if (wait) runBlocking { waitUntilRestOrThrow() }
-
-            runComposeProjectCommand("images").waitFor(10, TimeUnit.SECONDS)
-        }
-
-        @JvmStatic
-        fun stopCompose() {
-            val cordaStopProcess = runComposeProjectCommand("down")
-            cordaStopProcess.waitFor(30, TimeUnit.SECONDS)
-        }
-    }
 
     @field:TempDir
     lateinit var projectDir: File
-    protected lateinit var buildFile: File
+    private lateinit var buildFile: File
     private lateinit var networkPath: String
+
+    private val artifactoryUsernameString = "findProperty('cordaArtifactoryUsername') ?: System.getenv('CORDA_ARTIFACTORY_USERNAME')"
+    private val artifactoryPasswordString = "findProperty('cordaArtifactoryPassword') ?: System.getenv('CORDA_ARTIFACTORY_PASSWORD')"
 
     @BeforeEach
     fun setup() {
@@ -154,8 +79,8 @@ abstract class SmokeTestBase {
                             basic(BasicAuthentication)
                         }
                         credentials {
-                            username = findProperty('cordaArtifactoryUsername') ?: System.getenv('CORDA_ARTIFACTORY_USERNAME')
-                            password = findProperty('cordaArtifactoryPassword') ?: System.getenv('CORDA_ARTIFACTORY_PASSWORD')
+                            username = $artifactoryUsernameString
+                            password = $artifactoryPasswordString
                         }
                     }
                 }
@@ -168,22 +93,18 @@ abstract class SmokeTestBase {
 
     private fun appendArtifactoryCredentialsToTheCordaRuntimeGradlePluginExtension() {
         val existingContent = buildFile.readText()
+        val cordaRuntimeGradlePluginBlockCredentialsHeader = """
+             cordaRuntimeGradlePlugin {
+                artifactoryUsername = $artifactoryUsernameString
+                artifactoryPassword = $artifactoryPasswordString
+        """.trimIndent()
         val newContent = if (existingContent.contains("cordaRuntimeGradlePlugin")) {
             existingContent.replace(
                 "cordaRuntimeGradlePlugin {",
-                """
-                     cordaRuntimeGradlePlugin {
-                        artifactoryUsername = findProperty('cordaArtifactoryUsername') ?: System.getenv('CORDA_ARTIFACTORY_USERNAME')
-                        artifactoryPassword = findProperty('cordaArtifactoryPassword') ?: System.getenv('CORDA_ARTIFACTORY_PASSWORD')
-                """.trimIndent()
+                cordaRuntimeGradlePluginBlockCredentialsHeader
             )
         } else {
-            existingContent + """
-                cordaRuntimeGradlePlugin {
-                    artifactoryUsername = findProperty('cordaArtifactoryUsername') ?: System.getenv('CORDA_ARTIFACTORY_USERNAME')
-                    artifactoryPassword = findProperty('cordaArtifactoryPassword') ?: System.getenv('CORDA_ARTIFACTORY_PASSWORD')
-                }
-            """.trimIndent()
+            listOf(existingContent, cordaRuntimeGradlePluginBlockCredentialsHeader, "}").joinToString("\n")
         }
         buildFile.writeText(newContent)
     }
